// frontend/src/types/textProps.ts
export interface HeaderTextProps {
  home: string;
  create: string;
  myLeagues: string;
  signIn: string;
  signUp: string;
  logOut: string;
  close: string;
}
export interface HomeTextProps {
  homeGreeting: string;
  createPage: string;
  myPages: string;
  join: string;
  league: string;
  signIn: string;
  signUp: string;
  startCreating: string;
  signInToCreate: string;
  easyManage: string;
}
export interface SigninTextProps {
  email: string;
  emailPlaceholder: string;
  password: string;
  passwordPlaceholder: string;
  signIn: string;
  signingIn: string;
  error: string;
  donotHaveAccount: string;
  createAccount: string;
}
export interface SignupTextProps {
  createAccount: string;
  alreadyHaveAccount: string;
  signIn: string;
  nameLabel: string;
  emailLabel: string;
  passwordLabel: string;
  confirmPasswordLabel: string;
  sendVerificationCode: string;
  verificationCodePlaceholder: string;
  verifyCode: string;
  signUpButton: string;
  signingUp: string;
  successMessage: string;
  errorMessage: string;
}
export interface MyLeaguesTextsProps {
  title: string;
  createNewLeague: string;
  fetchFailed: string;
}
export interface CreateLeagueTextProps {
  title: string;
  editTeamNames: string;
  save: string;
  cancel: string;
  team: string;
  leagueName: string;
  numberOfTeamsText: string;
  playersPerTeam: string;
  matchDuration: string;
  breakDuration: string;
  totalPlayTime: string;
  numberOfPlaygrounds: string;
  startDate: string;
  startTime: string;
  register: string;
  sunday: string;
  monday: string;
  tuesday: string;
  wednesday: string;
  thursday: string;
  friday: string;
  saturday: string;
}

export interface LeagueTextProps {
  joinTeamNow: string;
  ifTeamsFull: string;
  tryAgain: string,
  selectPosition: string;
  enterName: string;
  registering: string;
  register: string;
  edit: string;
  ok: string;
  cancel: string;
  delete: string;
  confirmDelete: string;
  share: string;
  matchesSchedule: string;
  viewMatches: string;
  goalkeeper: string;
  defender: string;
  midfielder: string;
  forward: string;
}